+++ src/aspect.js
@@ -41,6 +41,40 @@ export let aspect = ({
   return x[name]
 }
 
+export let aspectSync = ({
+  name = 'aspect',
+  init = _.noop,
+  after = _.noop,
+  before = _.noop,
+  always = _.noop,
+  onError = throws
+  // ?: interceptParams, interceptResult, wrap
+}) => f => {
+  let {state = {}} = f
+  init(state)
+  // Trick to set function.name of anonymous function
+  let x = {
+    [name]: (...args) => {
+      let result
+      let error
+      try {
+        before(args, state)
+        result = f(...args)
+        after(result, state, args)
+      } catch (e) {
+        error = onError(e, state, args)
+      } finally {
+        always(state, args)
+        if (error)
+          throw error
+        return result
+      }
+    }
+  }
+  x[name].state = state
+  return x[name]
+}
+
 // Example Aspects
 let logs = (extend = defaultsOn) => aspect({
   init: extend({ logs: [] }),
@@ -109,11 +143,17 @@ let command = (extend, timeout) => _.flow(
   error(extend)
 )
 
+let deprecate = (subject, version, alternative) => aspectSync({
+  before: () =>
+    console.warn(`${subject} is deprecated${version ? ` as of ${version}`: ''}${alternative ? ` in favor of ${alternative}`: ''}`)
+})
+
 export let aspects = {
   logs,
   error,
   errors,
   status,
+  deprecate,
   clearStatus,
   concurrency,
   command